refactored to make parameters be doubles instead of ints

+++ src/main/java/app/services/SvgService.java
@@ -4,9 +4,9 @@ public interface SvgService
 {
     void addRectangle(double x, double y, double height, double width, String style);
 
-    void addLine(int x1, int y1, int x2, int y2, String style);
+    void addLine(double x1, double y1, double x2, double y2, String style);
 
-    void addArrow(int x1, int y1, int x2, int y2, String style);
+    void addArrow(double x1, double y1, double x2, double y2, String style);
 
     void addText(int x, int y, int rotation, String text);
